rename helper/ to helpers/
