+++ app/admin/dashboard/article/page.tsx
@@ -5,7 +5,7 @@ import { PiArticleMedium } from "react-icons/pi";
 import { useEffect, useState } from "react";
 import { CiTrash } from "react-icons/ci";
 import { MdOutlineModeEdit } from "react-icons/md";
-import { timeFormatter } from "@/helper/timeFormatterToID";
+import { timeFormatter } from "@/helpers/timeFormatterToID";
 import { RiExpandDiagonalLine } from "react-icons/ri";
 import { useRouter } from "next/navigation";
