+++ app/api/article/id/[id]/route.ts
@@ -1,9 +1,9 @@
 import prisma from "@/libs/prisma";
 import { Prisma } from "@/generated/prisma/client";
 import * as z from "zod";
-import { validateBody } from "@/helper/requestHelper";
-import { validateJwtAuthHelper } from "@/helper/authHelper";
-import { generateSlug } from "@/helper/generateSlugHelper";
+import { validateBody } from "@/helpers/requestHelper";
+import { validateJwtAuthHelper } from "@/helpers/authHelper";
+import { generateSlug } from "@/helpers/generateSlugHelper";
 
 const ArticleSchema = z.object({
   title: z.string().min(5),
