+++ app/api/auth/login/route.ts
@@ -4,7 +4,7 @@ import { Prisma } from "@/generated/prisma/client";
 import * as z from "zod";
 import jwt from "jsonwebtoken";
 import { AUTH_CONFIG } from "@/libs/config/JWTConfig";
-import { validateBody } from "@/helper/requestHelper";
+import { validateBody } from "@/helpers/requestHelper";
 
 const User = z.object({
   username: z.string().min(5),
