+++ app/api/auth/register/route.ts
@@ -2,7 +2,7 @@ import bcrypt from "bcryptjs";
 import prisma from "@/libs/prisma";
 import { Prisma } from "@/generated/prisma/client";
 import * as z from "zod";
-import { validateBody } from "@/helper/requestHelper";
+import { validateBody } from "@/helpers/requestHelper";
 
 // zod type validation
 const fromRequest = z.object({
