+++ app/api/shopitem/route.ts
@@ -1,10 +1,10 @@
-import { validateBody } from "@/helper/requestHelper";
-import { validateJwtAuthHelper } from "@/helper/authHelper";
+import { validateBody } from "@/helpers/requestHelper";
+import { validateJwtAuthHelper } from "@/helpers/authHelper";
 import * as z from "zod";
 import { JwtPayload } from "jsonwebtoken";
 import prisma from "@/libs/prisma";
 import { Prisma } from "@/generated/prisma/client";
-import { generateSlug } from "@/helper/generateSlugHelper";
+import { generateSlug } from "@/helpers/generateSlugHelper";
 
 const ShopItem = z.object({
   name: z.string(),
