+++ helpers/generateSlugHelper.ts
