+++ helpers/strValidatorHelper.ts
