+++ helpers/timeFormatterToID.ts
